refactor(29945): fix dataOnly for mocks
